Move shared HullWhite model code
