+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/model/HullWhiteOneFactorPiecewiseConstantParametersProviderTest.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.pricer.index;
+package com.opengamma.strata.pricer.model;
 
 import static com.opengamma.strata.basics.date.DayCounts.ACT_360;
 import static com.opengamma.strata.basics.date.DayCounts.ACT_ACT_ISDA;
